auth_service/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
PyMySQL==1.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
requests==2.32.3
argon2-cffi==23.1.0
